requirements.txt
discord.py[voice]
PyNaCl==1.3.0
async-timeout==3.0.1
youtube_dl
datetime
os
json
random
beautifulsoup4
requests